+++ 1.20/Common/src/main/java/fuzs/puzzleslib/api/core/v1/CommonAbstractions.java
@@ -1,18 +1,18 @@
 package fuzs.puzzleslib.api.core.v1;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.MenuProvider;
 import net.minecraft.world.damagesource.DamageSource;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.*;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.ServerLevelAccessor;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 
@@ -105,4 +105,14 @@ public interface CommonAbstractions {
      * @param interactionHand the hand holding the destroyed stack
      */
     void onPlayerDestroyItem(Player player, ItemStack itemStack, @Nullable InteractionHand interactionHand);
+
+    /**
+     * Retrieves a {@link MobSpawnType} from a {@link Mob} if it has been set during {@link Mob#finalizeSpawn(ServerLevelAccessor, DifficultyInstance, MobSpawnType, SpawnGroupData, CompoundTag)}.
+     * <p>Note that the spawn type is saved with the mob, so it persists across chunk and level reloads.
+     *
+     * @param mob the mob
+     * @return the spawn type or null if none has been set
+     */
+    @Nullable
+    MobSpawnType getMobSpawnType(Mob mob);
 }